% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tweet-thread.R
\name{as_tweet_thread}
\alias{as_tweet_thread}
\title{Turn a data frame of `rtweet` tweets into a tweet thread}
\usage{
as_tweet_thread(x)
}
\arguments{
\item{x}{a data frame of `rtweet` tweets}
}
\description{
A tweet thread is usually a series of tweets that are in reply to each other
from the same author. It may be advantageous to mark a series of tweets as
a tweet thread for display purposes.
}
