% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-thread.R
\name{get_thread}
\alias{get_thread}
\title{Retrieve an authors tweet thread/rant from a starting status id}
\usage{
get_thread(first_status, .timeline_history = 3000)
}
\arguments{
\item{first_status}{the first status id in the tweet thread}

\item{.timeline_history}{how much timeline history of the tweet author
to pull to use when searching for the thread?}
}
\value{
`list` with a data frame of the tweets and a text representation of the
        thread.
}
\description{
Retrieve an authors tweet thread/rant from a starting status id
}
\examples{
\dontrun{
thrd <- get_thread("988895496381984770")

thrd
}
}
