% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlify-tweet.R
\name{htmlify_tweet}
\alias{htmlify_tweet}
\title{Make a Tweet into an HTML object}
\usage{
htmlify_tweet(x, avatar = FALSE, images = FALSE)
}
\arguments{
\item{x}{one row from a rtweet data frame}

\item{avatar}{display avatars}

\item{images}{display images}
}
\value{
HTML representation of the tweet (character)
}
\description{
Make a Tweet into an HTML object
}
