% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-thread.R
\name{print.tweet_thread}
\alias{print.tweet_thread}
\title{Print method for a tweet thread}
\usage{
\method{print}{tweet_thread}(x, html = FALSE, avatar = TRUE,
  images = TRUE, css = NULL, theme = c("light", "dark"), ...)
}
\arguments{
\item{x}{thee tweet thread object}

\item{html}{print the HTML or text version of the tweet}

\item{avatar, images}{include avatars/images?}

\item{css}{custom CSS for the HTML tweet thread}

\item{theme}{dark or light theme?}

\item{...}{unused}
}
\value{
the tweet thread object (invisibly)
}
\description{
Print method for a tweet thread
}
