% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-tweet.R
\name{style_tweet}
\alias{style_tweet}
\title{Style a tweet with HTML}
\usage{
style_tweet(x, avatar = FALSE, images = FALSE, css = NULL,
  theme = c("light", "dark"), print = TRUE)
}
\arguments{
\item{x}{tweet (1 row from an `rtweet` data frame)}

\item{avatar, images}{include avatars/images?}

\item{css}{your own CSS to use vs the default}

\item{theme}{light or dark theme?}

\item{print}{display the tweet after styling?}
}
\value{
a styled tweet (invisibly)
}
\description{
Style a tweet with HTML
}
