% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetview-package.R
\docType{package}
\name{tweetview}
\alias{tweetview}
\alias{tweetview-package}
\title{Tools to Organize and Visualize Tweets}
\description{
Tools to Organize and Visualize Tweets
}
\author{
Bob Rudis (bob@rud.is)
}
